// src/common/base-text.d.ts
export interface BaseText {
  originX: string;
  originY: string;
  top: number;
  left: number;
  width: number;
  height: number;
  fill: string;
  fontSize: number;
  angle: number;
  scaleX: number;
  scaleY: number;
  fontWeight: string;
  fontFamily: string;
  textAlign: string;
  fontStyle: string;
  stroke: string;
  strokeWidth: number;
  containerWidth: number;
  containerHeight: number;
}